Applied node's security to delNode

+++ bin/node_type/delNode.php
@@ -23,8 +23,6 @@ if (!defined("CMS_VERSION")) {
     die("");
 }
 
-//TODO: Securizate access
-
 if (!class_exists("commandDelNode")) {
 
     class commandDelNode extends driverCommand {
@@ -44,23 +42,46 @@ if (!class_exists("commandDelNode")) {
             } else {
                 try {
                     $def = driverCommand::run("getNodeTypeDef", $params);
-                    // Delete relations
-                    foreach($def["fields"] as $field) {
-                        if ($field["multi"]) {
-                            $table = '`node_relation_'.$params["nodetype"].'_'
-                                    .$field["name"].'_'.$field["type"].'`';
-                            $sql = "delete from $table where `type1` = ".$params["nid"];
-                            dbConn::Execute($sql);
-                        }
+                    $nodeAccess = 0;
+                    $nodeUser_owner = 0;
+                    $nodeGroup_owner = 0;
+                    $sql = "select `id`, `access`, `user_owner`, `group_owner` from `node_{$params["nodetype"]}` where `id` = " . $params["nid"];
+                    $q = dbConn::Execute($sql);
+                    if (!$q->EOF) {
+                        $nodeAccess = $q->fields["access"];
+                        $nodeUser_owner = $q->fields["user_owner"];
+                        $nodeGroup_owner = $q->fields["group_owner"];
                     }
-                    // Delete node
-                    $sql = "delete from `node_{$params["nodetype"]}` where `id` = ".$params["nid"];
-                    dbConn::Execute($sql);
-                    // Delete page
-                    driverCommand::run("delPage", array(
-                        'name' => "node_type_" . $params["nodetype"] . "_" . $params["nid"],
-                    ));
-                    $resp["ok"] = true;
+                    $usrGrps = driverUser::getGroupsID();
+                    $allowed = driverUser::secNodeCanDelete($nodeAccess, 
+                            $nodeUser_owner == driverUser::getID(), 
+                            array_search($nodeGroup_owner, $usrGrps) !== FALSE);
+                    if (!$allowed) {
+                        $allowed = driverUser::secNodeCanDelete($def["access"], 
+                            $def["user_owner"] == driverUser::getID(), 
+                            array_search($def["group_owner"], $usrGrps) !== FALSE);
+                    }
+                    if ($allowed) {
+                        // Delete relations
+                        foreach($def["fields"] as $field) {
+                            if ($field["multi"]) {
+                                $table = '`node_relation_'.$params["nodetype"].'_'
+                                        .$field["name"].'_'.$field["type"].'`';
+                                $sql = "delete from $table where `type1` = ".$params["nid"];
+                                dbConn::Execute($sql);
+                            }
+                        }
+                        // Delete node
+                        $sql = "delete from `node_{$params["nodetype"]}` where `id` = ".$params["nid"];
+                        dbConn::Execute($sql);
+                        // Delete page
+                        driverCommand::run("delPage", array(
+                            'name' => "node_type_" . $params["nodetype"] . "_" . $params["nid"],
+                        ));
+                        $resp["ok"] = true;
+                    } else {
+                        $resp["msg"] = "You can't delete nodes.";
+                    }
                 } catch (Exception $exc) {
                     $resp["msg"] = $exc->getMessage();
                 }
